birefnet: support flash attention in swin-v1
* redoing the measurements yields higher values than before for pytorch, not 100% sure if previous times were incorrect
* comparison was against official BiRefNet repo, a3bb3efe2f824ec66644ca5941583c4c90c6e027
* tried with SDPA on/off

+++ src/visp/arch/swin.h
@@ -24,7 +24,7 @@ struct layer_result {
 
 void compute_relative_position_index(span<int32_t> dst, int window_size);
 tensor_data create_relative_position_index(ggml_context* ctx, int window_size);
-void compute_attention_mask(std::span<float> out, int64_t w, int64_t h, int window_size);
+void compute_attention_mask(std::span<byte> out, int64_t w, int64_t h, int window_size);
 tensor_data create_attention_mask(ggml_context* ctx, int64_t w, int64_t h, int window_size);
 
 tensor mlp(model_ref m, tensor x);
